Integrate batch operations into existing azure-mgmt-resource package
- Moved batch models from separate _batch_models.py into main _models_py3.py
- Updated __init__.py imports to include BatchRequest, BatchRequests, BatchResponse, BatchResponseItem, BatchResponseStatus
- Removed separate _batch_models.py file to consolidate models
- Follows established ARM SDK patterns for resource management operations
- Integration maintains backwards compatibility with existing models
